+++ go-auth-service/migrations/004_create_password_reset_tokens.sql
@@ -1,13 +1,15 @@
 -- Create password reset tokens table
 CREATE TABLE IF NOT EXISTS password_reset_tokens (
-    id INT AUTO_INCREMENT PRIMARY KEY,
-    user_id INT NOT NULL,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
+    user_id BIGINT NOT NULL,
     token VARCHAR(255) NOT NULL UNIQUE,
     used BOOLEAN DEFAULT FALSE,
     expires_at TIMESTAMP NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+
     INDEX idx_token (token),
     INDEX idx_user_id (user_id),
     INDEX idx_expires_at (expires_at),
+
     FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
 );
